feat: 스케줄러를 이용하여 TaskStatus가 REQ인 TaskReq의 파일 하나를 읽고 처리하는 배치 작업 기능 추가

+++ src/main/java/backend/week8/domain/dadDetReport/entity/DadDetReport.java
@@ -4,7 +4,10 @@ import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
-import javax.persistence.*;
+import javax.persistence.Column;
+import javax.persistence.EmbeddedId;
+import javax.persistence.Entity;
+import javax.persistence.Table;
 
 @Table(name = "DAD_DET_REPORT")
 @Entity
@@ -19,7 +22,7 @@ public class DadDetReport {
 	@Column(name = "CLICKS")
 	private int clicks;
 	@Column(name = "AVERAGE_IMPRESSION_RANK")
-	private int averageImpressionRank;
+	private double averageImpressionRank;
 	@Column(name = "AVERAGE_CLICK_COST")
 	private int averageClickCost;
 	@Column(name = "ADVERTISING_COST")
